Add command line interface for setting stuff.
You can set the active effect and color, as used by the arduino
to run effects. You can also set (globally for now) the "total
dimming" (brightness), strobe, "voice" (mode), and "speed" (for
built-in modes) that are sent to the lights. The built in modes
are as follows:

   0-50: obey DMX brightness+RGB values
 51-100: hard-cut rainbow
101-150: fade rainbow
151-200: pulse color cycle (rainbow ish)
201-255: sound activated color cycle

In the first mode, the brightness channel is obeyed. In the other
modes, the speed channel is obeyed. Strobe works in any mode.